Moved pure role logic into role.go files.
